pyproject.toml
[project]
name = "codebase-analyzer"
version = "0.1.0"
description = "Test Task for Python developer"
authors = [
    {name = "Andrei Kukaleu",email = "andrei_kukalev@epam.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph (>=0.3.21,<0.4.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-core (>=0.3.49,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "GitPython (>=3.1.44,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "networkx (>=3.4.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
